move logic out of world_map
