+++ aimmo-game/simulation/simulation_runner.py
@@ -54,7 +54,28 @@ class SimulationRunner(object):
         with self._lock:
             self._update_effects()
             num_avatars = len(self.game_state.avatar_manager.active_avatars)
-            self.game_state.world_map.update(num_avatars, self.game_state)
+            self.update(num_avatars, self.game_state)
+
+    def update(self, num_avatars, game_state):
+        self._update_avatars(game_state)
+        self._update_map(num_avatars)
+
+    def _update_avatars(self, game_state):
+        self._apply_score()
+        PickupApplier().apply(game_state)
+
+    def _apply_score(self):
+        for cell in self.game_state.world_map.score_cells():
+            try:
+                cell.avatar.score += 1
+            except AttributeError:
+                pass
+
+    def _update_map(self, num_avatars):
+        context = MapContext(num_avatars=num_avatars)
+        MapExpander().update(self, context=context)
+        ScoreLocationUpdater().update(self, context=context)
+        PickupUpdater().update(self, context=context)
 
     def _mark_complete(self):
         self.communicator.mark_game_complete(data=self.game_state.serialise())
